+++ core/views/fixed_assets/fixed_asset_maintenance_views.py
@@ -8,16 +8,21 @@ from django.utils.decorators import method_decorator
 from django.shortcuts import get_object_or_404
 from core.models import (
     AssetMaintenance,
+    FixedAsset,
 
 )
+from core.validators import _api_auth_required, _child_owned_object_or_404
 
 @method_decorator(csrf_exempt, name="dispatch")
 class AssetMaintenanceListView(View):
 
     def get(self, request):
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
         asset_id = request.GET.get("asset")
 
-        qs = AssetMaintenance.objects.all().order_by("-date", "-id")
+        qs = AssetMaintenance.objects.filter(asset__owner=request.user).order_by("-date", "-id")
 
         if asset_id:
             qs = qs.filter(asset_id=asset_id)
@@ -27,7 +32,11 @@ class AssetMaintenanceListView(View):
         })
 
     def post(self, request):
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
         data = json.loads(request.body)
+        get_object_or_404(FixedAsset, pk=data["asset_id"], owner=request.user)
 
         item = AssetMaintenance.objects.create(
             asset_id=data["asset_id"],
@@ -43,7 +52,10 @@ class AssetMaintenanceListView(View):
 class AssetMaintenanceDetailView(View):
 
     def put(self, request, pk):
-        item = get_object_or_404(AssetMaintenance, pk=pk)
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
+        item = _child_owned_object_or_404(AssetMaintenance, pk, request, parent_field="asset")
 
         data = json.loads(request.body)
 
@@ -58,7 +70,10 @@ class AssetMaintenanceDetailView(View):
         return JsonResponse(item.to_dict())
 
     def delete(self, request, pk):
-        item = get_object_or_404(AssetMaintenance, pk=pk)
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
+        item = _child_owned_object_or_404(AssetMaintenance, pk, request, parent_field="asset")
         item.delete()
 
         return JsonResponse({"deleted": pk})
