+++ core/views/fixed_assets/fixed_asset_valuation_views.py
@@ -8,16 +8,21 @@ from django.utils.decorators import method_decorator
 from django.shortcuts import get_object_or_404
 from core.models import (
     AssetValuationHistory,
+    FixedAsset,
 
 )
+from core.validators import _api_auth_required, _child_owned_object_or_404
 
 @method_decorator(csrf_exempt, name="dispatch")
 class AssetValuationHistoryListView(View):
 
     def get(self, request):
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
         asset_id = request.GET.get("asset")
 
-        qs = AssetValuationHistory.objects.all().order_by(
+        qs = AssetValuationHistory.objects.filter(asset__owner=request.user).order_by(
             "-valuation_date",
             "-id",
         )
@@ -32,7 +37,11 @@ class AssetValuationHistoryListView(View):
         })
 
     def post(self, request):
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
         data = json.loads(request.body)
+        get_object_or_404(FixedAsset, pk=data["asset_id"], owner=request.user)
 
         item = AssetValuationHistory.objects.create(
             asset_id=data["asset_id"],
@@ -57,10 +66,10 @@ class AssetValuationHistoryListView(View):
 class AssetValuationHistoryDetailView(View):
 
     def put(self, request, pk):
-        item = get_object_or_404(
-            AssetValuationHistory,
-            pk=pk,
-        )
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
+        item = _child_owned_object_or_404(AssetValuationHistory, pk, request, parent_field="asset")
 
         data = json.loads(request.body)
 
@@ -86,10 +95,10 @@ class AssetValuationHistoryDetailView(View):
         return JsonResponse(item.to_dict())
 
     def delete(self, request, pk):
-        item = get_object_or_404(
-            AssetValuationHistory,
-            pk=pk,
-        )
+        auth_error = _api_auth_required(request)
+        if auth_error:
+            return auth_error
+        item = _child_owned_object_or_404(AssetValuationHistory, pk, request, parent_field="asset")
         item.delete()
 
         return JsonResponse({"deleted": pk})
